Clean up and move to abstract to make testing easier
